refactor: moved services to compose folder
