+++ src/backends/__init__.py
@@ -1,5 +1,5 @@
 from .backend_interface import BackendInterface, ServiceType
-from .compose_backend import ComposeBackend
-from .kubernetes_backend import KubernetesBackend
+from .compose.compose_backend import ComposeBackend
+from .kubernetes.kubernetes_backend import KubernetesBackend
 
 __all__ = [BackendInterface, ComposeBackend, KubernetesBackend, ServiceType]
