+++ src/backends/compose/services/base_service.py
