+++ src/backends/compose/services/grafana-provisioning/dashboards/10619_rev1.json
