+++ src/backends/compose/services/grafana-provisioning/dashboards/10619_rev2.json
