+++ src/backends/compose/services/grafana-provisioning/dashboards/bitcoin_rpc_scraper.json
