+++ src/backends/compose/services/grafana-provisioning/dashboards/dashboards.yml
